tag delete rule

+++ src/database/repository/Tag.repository.ts
@@ -1,7 +1,13 @@
 import tagModel from '../models/Tag'
 import { Tag, TagInput } from '../../graphql/types'
 
+import {RecipeRepository} from './'
+
 export class TagRepository {
+    recipeRepository: RecipeRepository
+    constructor(){
+        this.recipeRepository = new RecipeRepository()
+    }
 
     async getAll(): Promise<Tag[]> {
         const tags = await tagModel.find()
@@ -19,7 +25,11 @@ export class TagRepository {
     }
 
     async delete(id: String): Promise<Boolean> {
-        const success = await tagModel.deleteOne({ _id: id })
-        return success.acknowledged
+        const hasRecipeWith = await this.recipeRepository.existsByPrincipalTag(id)
+        if(!hasRecipeWith){
+            const success = await tagModel.deleteOne({ _id: id })
+            return success.acknowledged
+        }
+        return false
     }
 }
